Update Peggle.lua
Create Frame no longer works the way it used to.  It needs to inherit the Background template.  Went through and changed all of the Frames that were throwing errors from <frame>:SetBackdrop.  Example of what was changed:

before:  i = CreateFrame("Frame", "", a);
after:  i = CreateFrame("Frame", "", a, BackdropTemplateMixin and "BackdropTemplate");

After doing this, the addon seems to work again.